Resolve "Range size overflows int type"

Closes #37

See merge request hectorjsmith/excel-change-handler!45

+++ ExcelChangeHandler/Api/Config/Configuration.cs
@@ -4,6 +4,6 @@
     {
         public bool ChangeHandlingEnabled { get; set; } = true;
 
-        public int MaxMemorySize { get; set; } = 250_000;
+        public long MaxMemorySize { get; set; } = 250_000;
     }
 }
